Semanas/Semana1/Dia1/requirements.txt
# Dependências do Dia 1 - Semana 1
# Instalar com: pip install -r requirements.txt

# APIs de IA
groq>=0.4.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilitários
python-dotenv>=1.0.0

# Nota: Estas são as dependências mínimas para o Dia 1
# Mais dependências serão adicionadas conforme o plano avança

